Non-system includes
 - Change to use target-based includes instead of duplicating SYSTEM include paths
 - Removal of warnings
 - Include new SPICE version

+++ src/engine/openspaceengine.cpp
@@ -793,12 +793,8 @@ void OpenSpaceEngine::loadSingleAsset(const std::string& assetPath) {
                 progressInfo.progress = (*it)->progress();
 
                 if ((*it)->nTotalBytesIsKnown()) {
-                    progressInfo.currentSize = static_cast<uint64_t>(
-                        (*it)->nSynchronizedBytes()
-                    );
-                    progressInfo.totalSize = static_cast<uint64_t>(
-                        (*it)->nTotalBytes()
-                    );
+                    progressInfo.currentSize = (*it)->nSynchronizedBytes();
+                    progressInfo.totalSize = (*it)->nTotalBytes();
                 }
 
                 loading = true;
